+++ README.md
@@ -1,0 +1,24 @@
+# Diagram Editor
+
+This project provides a simple diagram editor built with PyQt5. The original single-file implementation was split into modules for better maintainability.
+
+## Requirements
+
+- Python 3.8+
+- PyQt5
+
+Install dependencies with:
+
+```bash
+pip install -r requirements.txt
+```
+
+## Running
+
+Run the editor using:
+
+```bash
+python main.py
+```
+
+The application allows creating nodes, connecting them with edges and exporting the diagram as image or PDF.
